Avoid destroying containers that still have children
When an event destroy a container that still have children, do not
automatically destroy it and all its children

e.g.
A Machine container has children of type Rank
Machile and Rank are created on an event ``init´´ and destroyed on
``finalize´´. When the first rank calls the do_end trigger on the
ContainerCreateTrigger for the Machine type, it will have information to
destroy also an event of type Rank named (e.g.) rank_0.
The parent Machine container, will still have children rank_1 to _7 that
have not called ``finalize´´ yet.

This commit changes the container_unique_names to a map so that every
container will map to a structure that has the parent name and the ammount
of children. When nchilds == 0, the container can be finished.