Missing feature: Define container 0 on the source
Container 0 is implicit to Pajé. The description language assumed every
defined toplevel container to be child of 0.
In some situations it may be necessary to define links on the 0 container
(e.g. hierarchy of Cluster->Machine->Rank; messages are between ranks but
can happen between different clusters). Simple solution would be to create
another faux-root container, but would not be elegant on the visualization

This change adds support to define a container type called ROOT as the
toplevel, thus the user can define link, state and event types (even
though event and state types make not much sense on 0 ...)